requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.3
MarkupSafe==2.1.2
requests==2.31.0
python-dotenv==1.0.0
Flask-HTTPAuth==4.8.0
Flask-CORS==4.0.0
gunicorn==21.2.0
psutil==5.9.5

# Testing dependencies
pytest==7.4.0
pytest-mock==3.14.0
beautifulsoup4==4.12.2

# SQLite is part of the Python standard library, no need to install separately
